requirements.txt
awsebcli==3.16.0
boto3==1.10.34
botocore==1.13.34
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
Django==2.2.7
django-storages==1.8
docutils==0.15.2
future==0.16.0
idna==2.7
jmespath==0.9.4
pathspec==0.5.9
Pillow==6.2.1
pypiwin32==223
python-dateutil==2.8.0
pytz==2019.3
pywin32==227
PyYAML==3.13
requests==2.20.1
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.7
